requirements.txt
valohai-utils
matplotlib
nvidia-pyindex
nvidia-tao
tqdm